+++ scm-unit-tests/t-exec.scm
@@ -85,4 +85,27 @@
         (reverse path))))
  '(connect talk1 disconnect talk2 talk1 disconnect)) ; Because continuation captured during before, thunk/after are called twice
 
+(define data '())
+(dynamic-wind
+  (lambda () (set! data (list "Invoked outer (before)")))
+  (lambda () (set! data (append data (let ((path '())
+      (c #f))
+  (let ((add (lambda (s)
+               (set! path (cons s path)))))
+    (dynamic-wind
+      (lambda () (add 'connect))
+      (lambda ()
+        (add (call-with-current-continuation
+               (lambda (c0)
+                 (set! c c0)
+                 'talk1))))
+      (lambda () (add 'disconnect)))
+    (if (< (length path) 4)
+        (c 'talk2)
+        (reverse path)))))))
+  (lambda () (set! data (append data "Invoked outer (after)"))))
+
+(assert/equal  data 
+              '("Invoked outer (before)" connect talk1 disconnect connect talk2 disconnect . "Invoked outer (after)"))
+
 (unit-test-handler-results)